push before game change

+++ assignment/final/messageSys.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 #include <stdio.h>
 
-int8_t encode(int8_t, int8_t);
+int encode(int, int);
 
-int8_t decodeX(int8_t);
+int decodeX(int);
 
-int8_t decodeY(int8_t);
+int decodeY(int);
 
 #endif /* MESSAGESYS_H */
